+++ README.md
@@ -1,2 +1,34 @@
 # Retail-Mart
 Retail web Application
+
+#Things Implemented
+
+1. Rest service.
+2. Entity added
+3. package structure added
+4. lombok integration
+
+#Pending Task
+
+1. Add exception handling.
+2. Add AOP for tracing method time with annonation
+3. Add kafka integration
+4. Add Rest Template integration and Feign with custom thread pool
+5. Add reading configuration from yml file
+5. Add Mono and flux integration
+6. Add compleatable future,@async annotation
+7. Add Trace id ,span id implementation
+8. Implement work flow activi 
+9. Add jenkins stage
+10. Docker file
+11. Jdbc template, Named Jdbc (return single value, List of object(string,customObjects), JPA integration implementation (embedded key)
+12. Add Junit testcases
+13. Integration test cases
+14. swagger integration.
+15. spring security
+16. Cors filtering
+17. ssl configuration
+18. jmeter scripts
+19. add spring profile 
+
+
