--- src/main/java/com/sampleProject/RetailMart/service/LoginService.java
@@ -1,10 +1,0 @@
-package com.sampleProject.RetailMart.service;
-
-import com.sampleProject.RetailMart.dto.LoginDto;
-import com.sampleProject.RetailMart.model.Login;
-
-public interface LoginService {
-	public Login signUp(LoginDto login);
-	public String login(LoginDto login);
-	
-}
