+++ src/main/java/com/sampleproject/retailmart/RetailMartApplication.java
@@ -1,4 +1,4 @@
-package com.sampleProject.RetailMart;
+package com.sampleproject.retailmart;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
