+++ src/main/java/com/sampleproject/retailmart/configuration/RetailMartConfig.java
@@ -1,4 +1,4 @@
-package com.sampleProject.RetailMart.configuration;
+package com.sampleproject.retailmart.configuration;
 
 
 import java.util.Collections;
@@ -21,7 +21,7 @@ public class RetailMartConfig {
     public Docket api() { 
         return new Docket(DocumentationType.SWAGGER_2)  
           .select()                                  
-          .apis(RequestHandlerSelectors.basePackage("com.sampleProject.RetailMart"))              
+          .apis(RequestHandlerSelectors.basePackage("com.sampleproject.retailmart"))              
           .paths(PathSelectors.any()) 
           .build().apiInfo(apiInfo());                                           
     }
