+++ src/main/java/com/sampleproject/retailmart/repository/LoginRepository.java
@@ -1,4 +1,4 @@
-package com.sampleProject.RetailMart.repository;
+package com.sampleproject.retailmart.repository;
 
 import java.util.Optional;
 
@@ -6,7 +6,7 @@ import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.sampleProject.RetailMart.model.Login;
+import com.sampleproject.retailmart.entity.Login;
 
 @Repository
 public interface LoginRepository extends JpaRepository<Login, Long > {
